replaced all individual card move test files with a single card-moves.test.js suite
